@@ -4,74 +4,24 @@
   "author": "jw",
   "private": true,
   "scripts": {
-    "build": "vue-cli-service build",
-    "build-graphql": "babel server -d dist/server ",
-    "lint": "vue-cli-service lint",
     "start": "vue-cli-service serve",
-    "graphql": "cross-env NODE_ENV=development nodemon server/start.js"
+    "build": "vue-cli-service build",
+    "lint": "vue-cli-service lint"
   },
   "dependencies": {
-    "@babel/cli": "7.0.0",
-    "@babel/core": "^7.11.6",
-    "@babel/plugin-transform-runtime": "^7.11.5",
-    "@babel/polyfill": "^7.11.5",
-    "@vue/composition-api": "^1.0.0-beta.11",
-    "babel-plugin-transform-object-rest-spread": "^6.26.0",
-    "accounting": "^0.4.1",
-    "apollo-boost": "^0.4.9",
-    "apollo-cache-inmemory": "^1.6.6",
-    "apollo-client": "^2.6.10",
-    "apollo-link": "^1.2.14",
-    "apollo-link-error": "^1.1.13",
-    "apollo-link-http": "^1.5.17",
-    "axios": "^0.19.0",
-    "babel-core": "^7.0.0-bridge.0",
-    "babel-node": "^0.0.1-security",
-    "babel-polyfill": "^6.26.0",
-    "babel-register": "^6.26.0",
+    "ant-design-vue": "^2.0.0-beta.10",
+    "axios": "^0.20.0",
+    "classnames": "^2.2.6",
     "connect-history-api-fallback": "^1.6.0",
     "core-js": "^3.6.5",
-    "cors": "^2.8.5",
-    "countries-list": "^2.4.3",
-    "dayjs": "^1.8.16",
-    "dexie": "^2.0.4",
-    "dotenv": "^8.1.0",
-    "element-ui": "^2.4.5",
-    "enum": "^2.5.0",
-    "express": "^4.17.1",
-    "express-graphql": "^0.11.0",
-    "express-sse": "^0.5.1",
-    "global": "^4.4.0",
-    "graphql": "^15.3.0",
-    "graphql-tag": "^2.11.0",
-    "hotkeys-js": "^3.6.14",
-    "http-proxy-middleware": "^0.20.0",
-    "humps": "^2.0.1",
-    "keymaster": "^1.6.2",
-    "lodash": "^4.17.15",
-    "lowdb": "^1.0.0",
-    "md5.js": "^1.3.5",
-    "mkdirp": "^0.5.1",
-    "mockjs": "^1.0.1-beta3",
-    "net": "^1.0.2",
-    "node-fetch": "^2.6.1",
-    "number-precision": "^1.3.1",
-    "qs": "^6.9.4",
-    "ramda": "^0.26.1",
-    "regenerator-runtime": "^0.13.7",
-    "screenfull": "^4.2.1",
-    "shortid": "^2.2.8",
-    "vue": "^2.6.11",
-    "vue-apollo": "^3.0.4",
-    "vue-awesome-swiper": "^3.1.3",
-    "vue-i18n": "^8.13.0",
-    "vue-router": "^3.2.0",
-    "vuex": "^3.4.0",
-    "babel-preset-env": "^1.7.0",
-    "vuex-persistedstate": "^2.5.4"
+    "dotenv": "^8.2.0",
+    "ramda": "^0.27.1",
+    "vue": "^3.0.0-0",
+    "vue-router": "^4.0.0-beta.13",
+    "vuex": "^4.0.0-beta.4"
   },
   "devDependencies": {
-    "@babel/preset-flow": "^7.10.4",
+    "@vue/babel-plugin-jsx": "^1.0.0-rc.3",
     "@vue/babel-preset-jsx": "^1.1.2",
     "@vue/babel-sugar-inject-h": "^1.1.2",
     "@vue/cli-plugin-babel": "~4.5.0",
