@@ -1,8 +1,16 @@
+import { RouterView } from "vue-router";
+import { Spin } from "ant-design-vue";
+import { useStore } from "vuex";
+
 const App = () => {
+  const {
+    state: { spinning }
+  } = useStore();
+
   return (
-    <div id="app">
-      <RouterView></RouterView>
-    </div>
+    <Spin tip="Loading..." size="large" spinning={spinning}>
+      <RouterView style={{ textAlign: "center" }} />
+    </Spin>
   );
 };
 
