@@ -1,30 +1,20 @@
-import Vue from "vue";
-import VueRouter from "vue-router";
-
-// 解决ElementUI导航栏中的vue-router在3.0版本以上重复点菜单报错问题
-const originalPush = VueRouter.prototype.push;
-VueRouter.prototype.push = function push(location) {
-  return originalPush.call(this, location).catch(err => err);
-};
-
-Vue.use(VueRouter);
+import { createRouter, createWebHistory } from "vue-router";
 
 const routes = [
   {
     path: "/",
     name: "home",
     component: () => import(/* webpackChunkName: "home" */ "../router/Home")
-  },
-  {
-    path: "/home",
-    name: "home",
-    component: () => import(/* webpackChunkName: "home" */ "../router/Home")
   }
+  // {
+  // 	path: "/home",
+  // 	name: "home",
+  // 	component: () => import(/* webpackChunkName: "home" */ "../router/Home")
+  // }
 ];
 
-const router = new VueRouter({
-  mode: "history",
-  base: process.env.BASE_URL,
+const router = createRouter({
+  history: createWebHistory(process.env.BASE_URL),
   routes
 });
 
