@@ -1,11 +1,16 @@
-import Vue from "vue";
-import Vuex from "vuex";
+import { createStore } from "vuex";
 
-Vue.use(Vuex);
-
-export default new Vuex.Store({
-  state: {},
-  mutations: {},
+const store = createStore({
+  state: {
+    spinning: false
+  },
+  mutations: {
+    changeSpinning(state) {
+      state.spinning = !state.spinning;
+    }
+  },
   actions: {},
   modules: {}
 });
+
+export default store;
