@@ -27,25 +27,8 @@ module.exports = {
     resolve: {
       alias: {
         components: path.join(__dirname, "src/components"),
-        router: path.join(__dirname, "src/router"),
-        core: path.join(__dirname, "src/core"),
-        gql: path.join(__dirname, "src/gql"),
-        config: path.join(__dirname, "src/config")
+        router: path.join(__dirname, "src/router")
       }
     }
-  },
-  chainWebpack: config => {
-    config.module
-      .rule("vue")
-      .use("vue-loader")
-      .loader("vue-loader")
-      .tap(options => {
-        options.transpileOptions = {
-          transforms: {
-            dangerousTaggedTemplateString: true
-          }
-        };
-        return options;
-      });
   }
 };
